requirements.txt
# Core dependencies
google-generativeai>=0.3.0
huggingface-hub>=0.19.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0

# Quantization tools
optimum>=1.14.0
auto-gptq>=0.5.0
bitsandbytes>=0.41.0

# Data handling and storage
sqlite3
datasets>=2.14.0
safetensors>=0.4.0

# CLI and interaction
click>=8.0.0
rich>=13.0.0
tqdm>=4.64.0

# Utilities
pydantic>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
psutil>=5.9.0

# API Server dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0

# Optional inference tools
# llama-cpp-python>=0.2.0  # Uncomment if needed for GGUF inference